Created new camera file for handling camera actions
